backend/docker-images/db-writer-client/requirements.txt
pika==0.13.1
neo4j-driver
requests
prometheus_client